requirements.txt
absl-py==2.3.0
affine==2.4.0
ai4ts==0.0.3
asttokens==3.0.0
attrs==25.3.0
benchpots==0.4
branca==0.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
einops==0.8.1
executing==2.2.0
filelock==3.18.0
folium==0.19.5
fonttools==4.58.0
fsspec==2025.5.1
greenlet==3.2.2
grpcio==1.72.1
h5py==3.13.0
huggingface-hub==0.32.3
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lxml==5.4.0
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.5
OWSLib==0.33.0
packaging==25.0
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pillow==11.2.1
platformdirs==4.3.7
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
psycopg2==2.9.10
pure_eval==0.2.3
pyarrow==20.0.0
Pygments==2.19.1
pygrinder==0.7
pykalman==0.10.1
pyparsing==3.2.3
pypots==0.19
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
rasterio==1.4.3
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-base==0.12.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
setuptools==80.9.0
shapely==2.1.0
six==1.17.0
SQLAlchemy==2.0.40
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
tsdb==0.7.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
xyzservices==2025.4.0
